testing servo, display finished

+++ modules/wipers/wipers.h
@@ -11,7 +11,7 @@
 
 void wipersInit();
 void wipersUpdate();
-float wipersModeRead();
+int wipersModeRead();
 
 
 //=====[#include guards - end]=================================================
